KS-1-4 refactor kosmos

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -1,13 +1,13 @@
 package org.example.lesson_1
 
 
-fun main(){
-    val length : Long = 40_868_600_000L
-    val age : Byte = 27
-    val partOfDay : Float = 0.075F
-    val second : Short = 6480
-    val partsOfYear : Float = 2.0547945205479453E-4F
-    val apogee : Int = 327000
+fun main() {
+    val length: Long = 40_868_600_000L
+    val age: Byte = 27
+    val partOfDay: Float = 0.075F
+    val second: Short = 6480
+    val partsOfYear: Double = 2.0547945205479453E-4
+    val apogee: Int = 327000
 
     println(length)
     println(age)
